requirements.txt
mlflow<3,>=2.2
numpy==1.21.4
pandas==1.5.3
psutil==5.9.5
scikit-learn==1.2.2
scipy==1.9.3
typing-extensions==4.6.1
ruamel.yaml==0.17.31
dvc
tqdm
python-dotenv
dvc[s3]
tensorflow-gpu==2.10